% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chrom_counts}
\alias{chrom_counts}
\title{Returns chromosome counts from Chromosome Counts Database API}
\usage{
chrom_counts(taxa, rank = c("genus", "family", "majorGroup"), full = FALSE,
  foptions = list())
}
\arguments{
\item{taxa}{Taxonomic name(s) to query. Can be either a single name, a vector of multiple names or a list. If supplying multiple names, these must all be of the same \code{rank}.}

\item{rank}{Rank to query. Must be either 'genus', 'family', or 'majorGroup'}

\item{full}{Logical. Whether to return full records. Defaults to \code{FALSE} which returns only partial records. Partial records includes the resolved name as well as the gametophytic (n) and sporophytic (2n) counts.}

\item{foptions}{additional options to be passed to \code{GET}}
}
\value{
A \code{data.frame} containing all records matched by query
}
\description{
This function calls the Chromsome Counts Database (CCDB) API and returns all counts for specified higher taxa.
}
\examples{
\dontrun{

## Get all counts for genus Castilleja
chrom_counts("Castilleja", "genus")

## Get all counts for both Castilleja and Lachemilla
chrom_counts(c("Castilleja", "Lachemilla"), "genus")


}
}

