% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chrom_counts_rank}
\alias{chrom_counts_rank}
\title{Returns chromosome counts from Chromosome Counts Database API}
\usage{
chrom_counts_rank(taxa, rank = c("genus", "family", "majorGroup"),
  full = FALSE, foptions = list())
}
\arguments{
\item{taxa}{Taxonomic rank to query. Only a single query can be made by this function.}

\item{rank}{Rank to query. Must be either 'genus', 'family', or 'majorGroup'}

\item{full}{Logical. Whether to return full records. Defaults to \code{FALSE} which returns only partial records. Partial records includes the resolved name as well as the gametophytic (n) and sporophytic (2n) counts.}

\item{foptions}{additional options to be passed to \code{GET}}
}
\value{
A \code{data.frame} containing all records matched by query
}
\description{
This function calls the Chromsome Counts Database (CCDB) API and returns all counts for specified higher taxa.
}
\examples{
\dontrun{

## Get all counts for genus Castilleja
chrom_counts_rank("Castilleja", "genus")

}
}

