% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chrome_bp}
\alias{chrome_bp}
\title{Human chromesome lengths in basepairs}
\format{An object of class \code{numeric} of length 22.}
\source{
\url{https://en.wikipedia.org/wiki/Human_genome}
}
\usage{
chrome_bp
}
\description{
A vector containing the lengths of the 22 human autosomes in basepairs.
}
\keyword{datasets}

