% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrome.R
\name{chrome_map}
\alias{chrome_map}
\title{Calculate shared and not shared lengths.}
\usage{
chrome_map(data, key, chromosome, bp_start, bp_end, threshold = 2e+05)
}
\arguments{
\item{data}{Dataframe with \code{key}, \code{chromosome}, \code{bp_start},
and \code{bp_end} columns. Each row describes an IBD segment.}

\item{key}{String identifying the column in \code{data} that uniquely
identifies the pair of individuals corresponding to each IBD segment.}

\item{chromosome}{String indentifying the column in \code{data} that
indicates the chromosome of each IBD segment.}

\item{bp_start}{String indentifying the column in \code{data} that
indicates the starting basepair of each IBD segment.}

\item{bp_end}{String indentifying the column in \code{data} that
indicates the ending basepair of each IBD segment.}

\item{threshold}{Numeric value, greater than or equal to 0 for which a gap
should be classified as a "merge."}
}
\description{
\code{chrome_map} returns a dataframe with shared and not shared regions.
}
\details{
Based on a set of IBD segments, \code{chrome_map} calculate the lengths (in
basepairs) of the share segments as well as any unshared areas in each
chromosome (i.e., gaps). In addition, the \code{threshold} parameter can be
used to identifiy areas that are small (which \code{chrome_map} will then
classify as "merge").
}
\examples{
data(test_ibd_segments)
chrome_map(test_ibd_segments, "result_id", "chromosome", "bp_start", "bp_end")
}

