% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromer-package.R
\docType{package}
\name{chromer-package}
\alias{chromer}
\alias{chromer-package}
\title{Chromer: Programmatic access to Chromosome Counts Database}
\description{
This package is a R based interface to the Chromosome Counts Database (CCDB). Currently, this consists of just one function \code{\link{chrom_counts}}, which queries the database using the taxa supplied. The function \code{\link{summarize_counts}} cleans and processes the data.
}
\references{
Rice, A., L. Glick, S. Abadi, M. Einhorn, N.M. Kopelman, A. Salman-Minkov, J. Mayzel, O. Chay, and I. Mayrose. 2014. The Chromosome Counts Database (CCDB) -- a community resource of plant chromosome numbers. New Phytologist doi:10.1111/nph.13191.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/chromer/}
  \item \url{https://github.com/ropensci/chromer}
  \item Report bugs at \url{https://github.com/ropensci/chromer/issues}
}

}
\author{
\strong{Maintainer}: Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})

Authors:
\itemize{
  \item Matthew Pennell \email{mwpennell@gmail.com} (\href{https://orcid.org/0000-0002-2886-3970}{ORCID})
  \item Paula Andrea Martinez \email{orchid00@gmail.com} (\href{https://orcid.org/0000-0002-8990-1985}{ORCID})
}

}
\keyword{internal}
