% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{chromer}
\alias{chromer}
\alias{chromer-chromer}
\alias{chromer-package}
\title{Chromer: Programmatic access to Chromosome Counts Database}
\description{
This package is a R based interface to the Chromosome Counts Database (CCDB). Currently, this consists of just one function \code{\link{chrom_counts}}, which queries the database using the taxa supplied. The function \code{\link{summarize_counts}} cleans and processes the data.
}
\references{
Rice, A., L. Glick, S. Abadi, M. Einhorn, N.M. Kopelman, A. Salman-Minkov, J. Mayzel, O. Chay, and I. Mayrose. 2014. The Chromosome Counts Database (CCDB) -- a community resource of plant chromosome numbers. New Phytologist doi:10.1111/nph.13191.
}

