% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_ibd_segments}
\alias{test_ibd_segments}
\title{Segments shared IBD (simulated).}
\format{An object of class \code{data.frame} with 31 rows and 8 columns.}
\usage{
test_ibd_segments
}
\description{
A dataset containing simulated output from ersa
(\url{https://github.com/rmunoz12/ersa}), where each row describes an IBD
segment. The variables are as follows:
}
\details{
\itemize{
 \item id. identifier for each row.
 \item result_id. indentifier for each pair of individuals.
 \item indv1, indv2. names of individuals.
 \item chromosome. chromosomes where each IBD segment is shared.
 \item bp_start. starting basepair for each IBD segment.
 \item bp_end. ending basepair for each IBD segment.
 \item length. length in cM -- values are completely meaningless in this
     example data.
}
}
\keyword{datasets}

